pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lead-scoring-openenv"
version = "0.1.0"
description = "OpenEnv-compatible lead scoring (isolated from main Meta pipeline)"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2",
  "pydantic>=2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lead_scoring*"]
